requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
jinja2==3.1.2
pillow==10.1.0
python-dotenv==1.0.0 